refactor(core): #789 hexToColor3 공용 추출(rule-of-three) + NOISE_CONTRACT 가드 확장
PR #788(#774) reviewer 비차단 권고 2건 후속 — 순수 리팩터/가드 (셰이더 GLSL·색상·로직 변경 0).

Part 1 — hexToColor3 공용 추출:
- 3곳(solar-system-scene / procedural-planet-shader / sun-shader)의 바이트 동일
  사설 hex→Color3 헬퍼(hexToColor3 ×2 + hexToRgb01 ×1)를 신규 color-utils.ts 의
  hexToColor3 단일 SSoT 로 통합. 3곳 import 교체(사설 선언 잔존 0).
- sun-shader 는 hexToRgb01 → hexToColor3 통일 + 미사용 Color3 babylon import 제거.

Part 2 — NOISE_CONTRACT_LINES 정적 계약 확장 (sun↔planet GLSL noise 동일성 가드):
- 4→12 라인(hash33 3줄 전체 + hash13 + valueNoise floor/fract/trilinear 본체 4줄).
- 각 라인이 SUN·PLANET 셰이더 양쪽 toContain 검증 — 실측으로 공유 확인된 라인만 포함.
- granulation 합성(fbm(p*granulationScale))은 sun-specific(planet 부재)이라 계약 제외.

검증: core 744 tests PASS + typecheck 0. 셰이더 출력 불변(동일 구현 추출).

Closes #789
