+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Semantic Versioning을 따른다.
 
 ### Notes (infra/docs)
 
+- **[#789] sun-shader 후속 정비 — hexToColor3 공용 추출 + NOISE_CONTRACT 가드 확장** ([#789](https://github.com/coseo12/astro-simulator/issues/789)) — PR #788(#774) reviewer 비차단 권고 2건의 순수 리팩터/가드 후속 (셰이더 GLSL·색상·로직 변경 0). **rule-of-three 공용 추출**: 3곳(`solar-system-scene.ts` / `procedural-planet-shader.ts` / `sun-shader.ts`)의 바이트 동일 사설 hex→Color3 헬퍼(`hexToColor3` ×2 + `hexToRgb01` ×1)를 신규 [`packages/core/src/scene/color-utils.ts`](packages/core/src/scene/color-utils.ts)의 `hexToColor3` 단일 SSoT 로 통합, 3곳 import 교체(사설 선언 잔존 0). **noise 정적 계약 확장**: `sun-shader.test.ts`의 `NOISE_CONTRACT_LINES` 를 4→12 라인으로 확장(hash33 3줄 전체 + hash13 + valueNoise floor/fract/trilinear 본체 4줄). sun↔planet GLSL noise 텍스트 동일성 가드 강화 — 각 라인이 SUN·PLANET 셰이더 양쪽 `toContain` 검증. granulation 합성 라인은 sun-specific(planet 부재)이라 계약에서 제외(양쪽 동일성 계약). core 744 tests PASS + typecheck 0, 셰이더 출력 불변.
 - **[#795] 반복 운영 마찰 원인 박제 + pgrep self-match bracket fix** ([#795](https://github.com/coseo12/astro-simulator/issues/795)) — 매 세션·릴리스 반복되던 저비용 마찰 4건(squash auto-close 미발동 / `--delete-branch` worktree 충돌 / pgrep self-match 오탐 / concurrency CANCELLED 코스메틱)의 **구조 원인 + 표준 절차**를 [`docs/ops/operational-friction.md`](docs/ops/operational-friction.md) 에 박제 + CLAUDE.md 프로젝트 고유 보강 포인터. **pgrep/pkill self-match bracket fix**: `.claude/agents/qa.md` + CLAUDE.md 의 agent-browser 정리 명령을 `agent-browser-chrome-` → `agent-browser-chrome[-]` 로 정정(pkill 자기 셸 kill 위험 차단 = 안전 개선). hook `session-start-zombie-check.sh` 는 `grep -v` 로 이미 안전(변경 없음). agent-browser 정리 절차의 upstream 기여(Z-패턴 Phase 2)는 [#830](https://github.com/coseo12/astro-simulator/issues/830) 후속 분리(upstream 미보유 drift). qa.md 는 harness-managed 이나 기존 HARNESS-DRIFT 데코레이터(#309)가 drift 커버, CLAUDE.md 변경은 프로젝트 고유 보강 섹션(managed-block 외부).
 
 ## [0.48.0] — 2026-07-15
